+++ ShadowOnlyShader/Packages/com.hidano.shadow-only-shader/Runtime/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("com.hidano.shadow-only-shader.Editor")]
+[assembly: InternalsVisibleTo("com.hidano.shadow-only-shader.Tests.Editor")]